Remove phase, action and record endpoints
Closes #149
