+++ helerm/urls.py
@@ -2,16 +2,12 @@ from django.conf.urls import include, url
 from rest_framework.routers import DefaultRouter
 from helusers import admin
 
-from metarecord.views import (ActionViewSet, AttributeViewSet, ClassificationViewSet, FunctionViewSet, ExportView,
-                              PhaseViewSet, RecordViewSet, TemplateViewSet)
+from metarecord.views import AttributeViewSet, ClassificationViewSet, FunctionViewSet, ExportView, TemplateViewSet
 from users.views import UserViewSet
 
 
 router = DefaultRouter()
 router.register(r'function', FunctionViewSet)
-router.register(r'phase', PhaseViewSet)
-router.register(r'action', ActionViewSet)
-router.register(r'record', RecordViewSet)
 router.register(r'attribute', AttributeViewSet)
 router.register(r'template', TemplateViewSet, base_name='template')
 router.register(r'user', UserViewSet)
