+++ metarecord/views/__init__.py
@@ -1,7 +1,4 @@
 from .function import FunctionViewSet  # noqa
-from .phase import PhaseViewSet  # noqa
-from .action import ActionViewSet  # noqa
-from .record import RecordViewSet  # noqa
 from .attribute import AttributeViewSet  # noqa
 from .template import TemplateViewSet  # noqa
 from .export import ExportView  # noqa
